Rename layout files for OAM and HOT
