+++ src/features/layer_features_panel/layouts/hotProjectsLayout.ts
